+++ src/app/(pages)/layout.tsx
@@ -1,62 +1,16 @@
-'use client';
-
 import React from 'react';
 
-import { createTheme, ThemeProvider as MuiThemeProvider, CssBaseline, GlobalStyles } from '@mui/material';
-
+import { App } from '@/components';
+import { AppThemeProvider } from '@/contexts/ThemeContext';
 import '@/styles/globals.css';
-import { Header } from '@/components';
-import { AppThemeProvider, useTheme } from '@/contexts/ThemeContext';
-import { PyodideProvider } from '@/lib/pyodide';
 
-function App({ children }: { children: React.ReactNode }) {
-  const { themeMode } = useTheme();
+import type { Metadata } from 'next';
 
-  const theme = React.useMemo(
-    () =>
-      createTheme({
-        palette: {
-          mode: themeMode,
-        },
-        components: {
-          MuiCardHeader: {
-            styleOverrides: {
-              root: ({ theme }) => ({
-                backgroundColor:
-                  theme.palette.mode === 'light'
-                    ? theme.palette.grey[100]
-                    : theme.palette.grey[800],
-              }),
-            },
-          },
-        },
-      }),
-    [themeMode]
-  );
-
-  return (
-    <MuiThemeProvider theme={theme}>
-      <CssBaseline />
-      <GlobalStyles
-        styles={{
-          body: {
-            '--divider-color': theme.palette.divider,
-          },
-        }}
-      />
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'column',
-          height: '100vh',
-        }}
-      >
-        <Header />
-        <PyodideProvider>{children}</PyodideProvider>
-      </div>
-    </MuiThemeProvider>
-  );
-}
+// eslint-disable-next-line react-refresh/only-export-components
+export const metadata: Metadata = {
+  title: 'Moduloxs',
+  description: 'Block ML is a block programming environment for data analysis.',
+};
 
 export default function RootLayout({
   children,
