+++ src/components/index.tsx
@@ -1,5 +1,6 @@
 export { AboutButton } from './AboutButton';
 export { AboutModal } from './AboutModal';
+export { App } from './App';
 export { FileUploader } from './FileUploader';
 export { Header } from './Header';
 export { SearchBox } from './SearchBox';
